Refactor folder structure and rename classes
